Remove SOAP files and dependencies
