+++ composer.json
@@ -6,10 +6,8 @@
       "php": ">=5.6.0",
       "ext-curl": "*",
       "ext-json": "*",
-      "ext-soap": "*",
       "ext-mbstring": "*",
-      "guzzlehttp/guzzle":"^5|^6|^7",
-      "ext-openssl": "*"
+      "guzzlehttp/guzzle":"^5|^6|^7"
   },
     "require-dev": {
       "phpunit/phpunit": "~4.0",
