+++ init.php
@@ -34,38 +34,3 @@ require_once(__DIR__ . '/src/Onepay/Exceptions/RefundCreateException.php');
 require_once(__DIR__ . '/src/Onepay/Exceptions/SignException.php');
 require_once(__DIR__ . '/src/Onepay/Exceptions/TransactionCommitException.php');
 require_once(__DIR__ . '/src/Onepay/Exceptions/TransactionCreateException.php');
-
-// WEBPAY
-require_once(__DIR__ . '/src/Webpay/Configuration.php');
-require_once(__DIR__ . '/src/Webpay/Webpay.php');
-require_once(__DIR__ . '/src/Webpay/WebpayCapture.php');
-require_once(__DIR__ . '/src/Webpay/WebpayCompleteTransaction.php');
-require_once(__DIR__ . '/src/Webpay/WebPayMallNormal.php');
-require_once(__DIR__ . '/src/Webpay/WebPayNormal.php');
-require_once(__DIR__ . '/src/Webpay/WebpayNullify.php');
-require_once(__DIR__ . '/src/Webpay/WebpayOneClick.php');
-require_once(__DIR__ . '/src/Webpay/initTransactionResponse.php');
-require_once(__DIR__ . '/src/Webpay/wsInitTransactionOutput.php');
-require_once(__DIR__ . '/src/Webpay/getTransactionResult.php');
-require_once(__DIR__ . '/src/Webpay/getTransactionResultResponse.php');
-require_once(__DIR__ . '/src/Webpay/transactionResultOutput.php');
-require_once(__DIR__ . '/src/Webpay/cardDetail.php');
-require_once(__DIR__ . '/src/Webpay/wsTransactionDetailOutput.php');
-require_once(__DIR__ . '/src/Webpay/wsTransactionDetail.php');
-require_once(__DIR__ . '/src/Webpay/acknowledgeTransaction.php');
-require_once(__DIR__ . '/src/Webpay/acknowledgeTransactionResponse.php');
-require_once(__DIR__ . '/src/Webpay/initTransaction.php');
-require_once(__DIR__ . '/src/Webpay/wsInitTransactionInput.php');
-require_once(__DIR__ . '/src/Webpay/wpmDetailInput.php');
-require_once(__DIR__ . '/src/Webpay/nullificationInput.php');
-require_once(__DIR__ . '/src/Webpay/nullificationOutput.php');
-require_once(__DIR__ . '/src/Webpay/nullify.php');
-require_once(__DIR__ . '/src/Webpay/nullifyResponse.php');
-
-// SOAP
-require_once(__DIR__ . '/src/Webpay/SoapValidation.php');
-require_once(__DIR__ . '/src/Webpay/WSSESoap.php');
-require_once(__DIR__ . '/src/Webpay/WSSecuritySoapClient.php');
-require_once(__DIR__ . '/src/Webpay/XMLSecurityKey.php');
-require_once(__DIR__ . '/src/Webpay/XMLSecurityDSig.php');
-require_once(__DIR__ . '/src/Webpay/XMLSecEnc.php');
