+++ src/Webpay/Oneclick/MallInscription.php
@@ -7,30 +7,19 @@ use Transbank\Webpay\Oneclick;
 use Transbank\Webpay\Oneclick\Exceptions\InscriptionDeleteException;
 use Transbank\Webpay\Oneclick\Exceptions\InscriptionFinishException;
 use Transbank\Webpay\Oneclick\Exceptions\InscriptionStartException;
+use Transbank\Webpay\Oneclick\Responses\InscriptionFinishResponse;
+use Transbank\Webpay\Oneclick\Responses\InscriptionStartResponse;
+use Transbank\Webpay\Oneclick\Traits\NeedsConfiguration;
 
 class MallInscription
 {
+    use NeedsConfiguration;
+    
     const INSCRIPTION_START_ENDPOINT = 'rswebpaytransaction/api/oneclick/v1.0/inscriptions';
     const INSCRIPTION_FINISH_ENDPOINT = 'rswebpaytransaction/api/oneclick/v1.0/inscriptions/$TOKEN$';
     const INSCRIPTION_DELETE_ENDPOINT = 'rswebpaytransaction/api/oneclick/v1.0/inscriptions';
     
-    public static function getCommerceIdentifier($options)
-    {
-        if ($options == null) {
-            $commerceCode = Oneclick::getCommerceCode();
-            $apiKey = Oneclick::getApiKey();
-            $baseUrl = Oneclick::getIntegrationTypeUrl();
-        } else {
-            $commerceCode = $options->getCommerceCode();
-            $apiKey = $options->getApiKey();
-            $baseUrl = Oneclick::getIntegrationTypeUrl($options->getIntegrationType());
-        }
-        return array(
-            $commerceCode,
-            $apiKey,
-            $baseUrl,
-        );
-    }
+    
 
     public static function start($userName, $email, $responseUrl, $options = null)
     {
