--- src/Webpay/acknowledgeTransaction.php
@@ -1,7 +1,0 @@
-<?php
-namespace Transbank\Webpay;
-
-class acknowledgeTransaction
-{
-    public $tokenInput; //string
-}
