--- src/Webpay/baseBean.php
@@ -1,6 +1,0 @@
-<?php
-namespace Transbank\Webpay;
-
-class baseBean
-{
-}
