--- src/Webpay/cardDetail.php
@@ -1,8 +1,0 @@
-<?php
-namespace Transbank\Webpay;
-
-class cardDetail
-{
-    public $cardNumber; //string
-    public $cardExpirationDate; //string
-}
