--- src/Webpay/getTransactionResult.php
@@ -1,7 +1,0 @@
-<?php
-namespace Transbank\Webpay;
-
-class getTransactionResult
-{
-    public $tokenInput; //string
-}
