--- src/Webpay/getTransactionResultResponse.php
@@ -1,7 +1,0 @@
-<?php
-namespace Transbank\Webpay;
-
-class getTransactionResultResponse
-{
-    public $return; //transactionResultOutput
-}
