--- src/Webpay/initTransaction.php
@@ -1,7 +1,0 @@
-<?php
-namespace Transbank\Webpay;
-
-class initTransaction
-{
-    public $wsInitTransactionInput; //wsInitTransactionInput
-}
