--- src/Webpay/initTransactionResponse.php
@@ -1,7 +1,0 @@
-<?php
-namespace Transbank\Webpay;
-
-class initTransactionResponse
-{
-    public $return; //wsInitTransactionOutput
-}
