--- src/Webpay/nullify.php
@@ -1,7 +1,0 @@
-<?php
-namespace Transbank\Webpay;
-
-class nullify
-{
-    public $nullificationInput;
-}
