--- src/Webpay/nullifyResponse.php
@@ -1,7 +1,0 @@
-<?php
-namespace Transbank\Webpay;
-
-class nullifyResponse
-{
-    public $return; //nullificationOutput
-}
