--- src/Webpay/wsInitTransactionOutput.php
@@ -1,8 +1,0 @@
-<?php
-namespace Transbank\Webpay;
-
-class wsInitTransactionOutput
-{
-    public $token; //string
-    public $url; //string
-}
