--- src/Webpay/wsTransactionDetail.php
@@ -1,11 +1,0 @@
-<?php
-namespace Transbank\Webpay;
-
-class wsTransactionDetail
-{
-    public $sharesAmount; //decimal
-    public $sharesNumber; //int
-    public $amount; //decimal
-    public $commerceCode; //string
-    public $buyOrder; //string
-}
