--- src/Webpay/wsTransactionDetailOutput.php
@@ -1,9 +1,0 @@
-<?php
-namespace Transbank\Webpay;
-
-class wsTransactionDetailOutput
-{
-    public $authorizationCode; //string
-    public $paymentTypeCode; //string
-    public $responseCode; //int
-}
